backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
google-generativeai
python-dotenv
httpx
python-multipart